Adaptación para MULTIENTIDAD

+++ sql/tool/run-scripts-from-tag.sh
@@ -23,12 +23,22 @@ function print_banner() {
     echo "******************************************************************************************"
 }
 
+function getConnectionParam() {
+    filename=`basename $1`
+    schema=`echo $filename | cut -d_ -f3`
+    if [[ $schema == "BANKMASTER" ]] || [[ $schema == "MASTER" ]] || [[ $schema == "HAYAMASTER" ]]; then
+        echo "$3"
+    else
+        echo "$2/$3"
+    fi
+} 
+
 function registerSQLScript() {
     git log $1 >> /dev/null 2>&1
     if [ $? -eq 0 ]; then
         HASH=`git rev-list HEAD $1 | tail -n 1`
         DATE=`git show -s --format="%ct" $HASH --`
-        printf "%s#%s \n" "$DATE" $1 >> $2
+        printf "%s#%s %s\n" "$DATE" $1 $3 >> $2
     fi
 }
 
@@ -87,20 +97,69 @@ print_banner
 
 BASEDIR=$(dirname $0)
 
+export SETENVGLOBAL=~/setEnvGlobal.sh
+if [ -f ~/setEnvGlobal${CUSTOMER_IN_UPPERCASE}.sh ] ; then
+  export SETENVGLOBAL=~/setEnvGlobal${CUSTOMER_IN_UPPERCASE}.sh
+fi
+if [ ! -f $SETENVGLOBAL ]; then
+    echo "No existe el fichero: $SETENVGLOBAL"
+    echo "Consulta las plantillas que hay en sql/tool/templates"
+    exit 1
+fi
+source $SETENVGLOBAL
+
 rm -rf $BASEDIR/tmp/*.txt $BASEDIR/tmp/*.log $BASEDIR/tmp/*.sh $BASEDIR/tmp/*.sql
 
+#PRODUCTO
 for file in `git diff $1 --name-only sql/**/producto/*.sql`
 do
-    registerSQLScript $file $BASEDIR/tmp/product-list-from-tag.txt 
+    if [ "$MULTIENTIDAD" != "" ] ; then
+        IFS=',' read -a entidades <<< "$MULTIENTIDAD"
+        for entidad in "${entidades[@]}"
+        do
+            connectionParam=`getConnectionParam $file ${!entidad} $3`
+            registerSQLScript $file $BASEDIR/tmp/product-list-from-tag.txt $connectionParam
+        done
+    else
+        registerSQLScript $file $BASEDIR/tmp/product-list-from-tag.txt $3
+    fi
 done
 
+#CLIENTE
 for file in `git diff $1 --name-only sql/**/$CUSTOMER_IN_LOWERCASE/*.sql`
 do
-    registerSQLScript $file $BASEDIR/tmp/customer-list-from-tag.txt
+    if [ "$MULTIENTIDAD" != "" ] ; then
+        IFS=',' read -a entidades <<< "$MULTIENTIDAD"
+        for entidad in "${entidades[@]}"
+        do
+            connectionParam=`getConnectionParam $file ${!entidad} $3`
+            registerSQLScript $file $BASEDIR/tmp/customer-list-from-tag.txt $connectionParam
+        done
+    else
+        registerSQLScript $file $BASEDIR/tmp/customer-list-from-tag.txt $3
+    fi
 done
 
-cat $BASEDIR/tmp/product-list-from-tag.txt | sort | cut -d# -f2 > $BASEDIR/tmp/list-from-tag.txt
-cat $BASEDIR/tmp/customer-list-from-tag.txt | sort | cut -d# -f2 >> $BASEDIR/tmp/list-from-tag.txt
+#SUBCLIENTE EN CASO DE MULTIENTIDAD
+if [ "$MULTIENTIDAD" != "" ] ; then
+    IFS=',' read -a entidades <<< "$MULTIENTIDAD"
+    for entidad in "${entidades[@]}"
+    do
+        SUBENTITY=`echo $entidad | tr '[:upper:]' '[:lower:]'`
+        for file in `git diff $1 --name-only sql/**/$CUSTOMER_IN_LOWERCASE/$SUBENTITY/*.sql`
+        do
+            connectionParam=`getConnectionParam $file ${!entidad} $3`
+            registerSQLScript $file $BASEDIR/tmp/customer-list-from-tag.txt $connectionParam
+        done
+    done
+fi
+
+if [ -f $BASEDIR/tmp/product-list-from-tag.txt ] ; then
+    cat $BASEDIR/tmp/product-list-from-tag.txt | sort | cut -d# -f2 > $BASEDIR/tmp/list-from-tag.txt
+fi
+if [ -f $BASEDIR/tmp/customer-list-from-tag.txt ] ; then
+    cat $BASEDIR/tmp/customer-list-from-tag.txt | sort | cut -d# -f2 >> $BASEDIR/tmp/list-from-tag.txt
+fi
 
 
 if [[ "$#" -ge 4 ]] && [[ "$4" == "go!" ]]; then
@@ -108,11 +167,11 @@ if [[ "$#" -ge 4 ]] && [[ "$4" == "go!" ]]; then
     do
         if [[ "$5" == "-v" ]]; then
             echo "--------------------------------------------------------------------------------"
-            echo "$BASEDIR/run-single-script.sh $line $3 $CUSTOMER_IN_UPPERCASE -v"
-            $BASEDIR/run-single-script.sh $line $3 $CUSTOMER_IN_UPPERCASE -v
+            echo "$BASEDIR/run-single-script.sh $line $CUSTOMER_IN_UPPERCASE -v"
+            $BASEDIR/run-single-script.sh $line $CUSTOMER_IN_UPPERCASE -v
             echo "--------------------------------------------------------------------------------"
         else
-            $BASEDIR/run-single-script.sh $line $3 $CUSTOMER_IN_UPPERCASE
+            $BASEDIR/run-single-script.sh $line $CUSTOMER_IN_UPPERCASE
         fi
     done < $BASEDIR/tmp/list-from-tag.txt
 else
@@ -121,7 +180,7 @@ else
     echo ""
     while read -r line
     do
-        echo "$BASEDIR/run-single-script.sh $line $3 $CUSTOMER_IN_UPPERCASE"
+        echo "$BASEDIR/run-single-script.sh $line $CUSTOMER_IN_UPPERCASE"
     done < $BASEDIR/tmp/list-from-tag.txt
     echo ""
     echo "Si estás de acuerdo, añade go! al final de la línea de comandos"
